Struct paho_mqtt3as_sys::MQTTAsync_disconnectOptions

#[repr(C)]

pub struct MQTTAsync_disconnectOptions {
    pub struct_id: [c_char; 4],
    pub struct_version: c_int,
    pub timeout: c_int,
    pub onSuccess: MQTTAsync_onSuccess,
    pub onFailure: MQTTAsync_onFailure,
    pub context: *mut c_void,
}

Fields
struct_id: [c_char; 4]

The eyecatcher for this structure. Must be MQTD.

struct_version: c_int

The version number of this structure. Must be 0 or 1. 0 signifies no SSL options

timeout: c_int

The client delays disconnection for up to this time (in milliseconds) in order to allow in-flight message transfers to complete.

onSuccess: MQTTAsync_onSuccess

A pointer to a callback function to be called if the disconnect successfully completes. Can be set to NULL, in which case no indication of successful completion will be received.

onFailure: MQTTAsync_onFailure

A pointer to a callback function to be called if the disconnect fails. Can be set to NULL, in which case no indication of unsuccessful completion will be received.

context: *mut c_void

A pointer to any application-specific context. The the context pointer is passed to success or failure callback functions to provide access to the context information in the callback.