Struct paho_mqtt3as_sys::MQTTAsync_successData

#[repr(C)]

pub struct MQTTAsync_successData {
    pub token: MQTTAsync_token,
    pub alt: MQTTAsync_successData__bindgen_ty_1,
}

The data returned on completion of a successful API call in the response callback onSuccess.

Fields
token: MQTTAsync_token

A token identifying the successful request. Can be used to refer to the request later.

alt: MQTTAsync_successData__bindgen_ty_1