Struct paho_mqtt3as_sys::MQTTClient_persistence

#[repr(C)]

pub struct MQTTClient_persistence {
    pub context: *mut c_void,
    pub popen: Persistence_open,
    pub pclose: Persistence_close,
    pub pput: Persistence_put,
    pub pget: Persistence_get,
    pub premove: Persistence_remove,
    pub pkeys: Persistence_keys,
    pub pclear: Persistence_clear,
    pub pcontainskey: Persistence_containskey,
}

@brief A structure containing the function pointers to a persistence implementation and the context or state that will be shared across all the persistence functions.

Fields
context: *mut c_void

A pointer to any data required to initialize the persistent store.

popen: Persistence_open

A function pointer to an implementation of Persistence_open().

pclose: Persistence_close

A function pointer to an implementation of Persistence_close().

pput: Persistence_put

A function pointer to an implementation of Persistence_put().

pget: Persistence_get

A function pointer to an implementation of Persistence_get().

premove: Persistence_remove

A function pointer to an implementation of Persistence_remove().

pkeys: Persistence_keys

A function pointer to an implementation of Persistence_keys().

pclear: Persistence_clear

A function pointer to an implementation of Persistence_clear().

pcontainskey: Persistence_containskey

A function pointer to an implementation of Persistence_containskey().