Type Definition paho_mqtt3as_sys::MQTTAsync_onFailure

type MQTTAsync_onFailure = Option<unsafe extern "C" fn(_: *mut c_void, _: *mut MQTTAsync_failureData)>;

This is a callback function. The client application must provide an implementation of this function to enable asynchronous notification of the unsuccessful completion of an API call. The function is registered with the client library by passing it as an argument in ::MQTTAsync_responseOptions. @param context A pointer to the context value originally passed to ::MQTTAsync_responseOptions, which contains any application-specific context. @param response Any failure data associated with the API completion.